backend/pyproject.toml
[project]
name = "releasetracker"
version = "0.1.0"
license = {text = "GPL-3.0"}
description = "轻量级版本追踪软件"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.4",
    "python-multipart>=0.0.6",
    "PyJWT>=2.8.0",
    "passlib>=1.7.4",
    "bcrypt==3.2.2",
    "cryptography>=41.0.0",
    "emoji>=2.10.0",
    "authlib>=1.6.8",
    "docker",
    "podman",
    "kubernetes",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
